+++ app/lib/registration/sign_up_screen.dart
@@ -9,13 +9,13 @@ import 'package:air/navigation/navigation.dart';
 import 'package:air/theme/theme.dart';
 import 'package:air/ui/colors/themes.dart';
 import 'package:air/ui/components/desktop/width_constraints.dart';
+import 'package:air/ui/icons/app_icons.dart';
 import 'package:air/ui/typography/font_size.dart';
 import 'package:air/widgets/widgets.dart';
 import 'package:air/util/cached_memory_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:iconoir_flutter/iconoir_flutter.dart' as iconoir;
 import 'package:provider/provider.dart';
 
 import 'registration_cubit.dart';
@@ -206,13 +206,10 @@ class _UserAvatarPicker extends StatelessWidget {
                 color: colors.fill.tertiary,
               ),
               alignment: Alignment.center,
-              child: IgnorePointer(
+              child: const IgnorePointer(
                 child: IconTheme(
-                  data: const IconThemeData(),
-                  child: iconoir.MediaImagePlus(
-                    width: 24,
-                    color: colors.text.primary,
-                  ),
+                  data: IconThemeData(),
+                  child: AppIcon.imagePlus(size: 24),
                 ),
               ),
             ),